mesh nodes refining

+++ src/element.zig
@@ -4,19 +4,19 @@ pub const Element = union(enum) {
     quadrilateral: Quadrilateral,
     triangle: Triangle,
 
-    //pub fn move(self: Element, dx: f32, dy: f32) void {
+    //pub fn move(self: Element, dx: f64, dy: f32) void {
     //    switch (self) {
     //        inline else => |s| s.move(dx, dy),
     //    }
     //}
 
-    pub fn getXValues(self: Element) struct { xs: []const f32, len: usize } {
+    pub fn getXValues(self: Element) struct { xs: []const f64, len: usize } {
         return switch (self) {
             inline else => |s| .{ .xs = s.x[0..], .len = s.x.len },
         };
     }
 
-    pub fn getYValues(self: Element) struct { ys: []const f32, len: usize } {
+    pub fn getYValues(self: Element) struct { ys: []const f64, len: usize } {
         return switch (self) {
             inline else => |s| .{ .ys = s.y[0..], .len = s.y.len },
         };
@@ -30,8 +30,8 @@ pub const Hexahedran = struct {
     //  | 3------2
     //  |/     |/
     //  7------6
-    x: [8]f32,
-    y: [8]f32,
+    x: [8]f64,
+    y: [8]f64,
 };
 
 pub const Tetrahedraon = struct {
@@ -42,8 +42,8 @@ pub const Tetrahedraon = struct {
     //    \ | /
     //     \|/
     //      2
-    x: [4]f32,
-    y: [4]f32,
+    x: [4]f64,
+    y: [4]f64,
 };
 
 pub const Quadrilateral = struct {
@@ -53,8 +53,8 @@ pub const Quadrilateral = struct {
     //   3----____          \
     //            ----____   \
     //                    ----2
-    x: [4]f32,
-    y: [4]f32,
+    x: [4]f64,
+    y: [4]f64,
 };
 
 pub const Triangle = struct {
@@ -62,6 +62,6 @@ pub const Triangle = struct {
     //     / \
     //    /   \
     //   2-----1
-    x: [3]f32,
-    y: [3]f32,
+    x: [3]f64,
+    y: [3]f64,
 };
